Fix spacing after \XMLfile with \xspace
While reading the documentation, I noticed that the phrase "XML file"
would not be separated from the next word by a space.  For instance, in
the abstract, the second to last sentence is like this:

So it also creates the XML filefor

where the words "file" and "for" are not separated by a space.  This can
be fixed by using the `\xspace` command when defining the `\XMLfile`
command as is implemented in this patch.